chatbot/requirements.txt
streamlit
langchain
langgraph
langchain.core
pydantic
langchain_google_genai
python-dotenv
google-generativeai
langchain-mcp-adapters
langchain_chroma
aiosqlite
pip install fastapi uvicorn